+++ src/components/Home/Confirm/CommentDetail/index.ts
@@ -3,3 +3,4 @@ export { default as CommentDetailInput } from './CommentDetailInput';
 export { default as CommentItem } from './CommentItem';
 export { default as CommentsInput } from './CommentsInput';
 export { default as CommentsItem } from './CommentsItem';
+export { default as NestedComments } from './NestedComments';
